enterprise-rag-platform/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.10.3
openai==1.57.3
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
chromadb==0.5.11
pypdf==4.3.1
sentence-transformers==3.0.1
